Create check idempotent matrix
Given a matrix, find whether it is Idempotent matrix or not. A matrix is said to be Idempotent if multiplied by itself returns the same matrix. The matrix M is said to be idempotent matrix if and only if M * M = M. In idempotent matrix M is a square matrix.